requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0
slowapi>=0.1.9
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
PyMuPDF>=1.23.8
openai>=1.12.0
httpx>=0.26.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
respx>=0.20.0

# Framework integrations
crewai>=0.19.0
pyautogen>=0.7.5
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0

# New dependencies
nest-asyncio>=1.6.0
